+++ src/Sql/Sql/ManagedInstance/Services/AzureSqlManagedInstanceAdapter.cs
@@ -177,7 +177,8 @@ namespace Microsoft.Azure.Commands.Sql.ManagedInstance.Adapter
                 DatabaseFormat = model.DatabaseFormat,
                 PricingModel = model.PricingModel,
                 IsGeneralPurposeV2 = model.IsGeneralPurposeV2,
-                StorageIOps = model.StorageIOps
+                StorageIOps = model.StorageIOps,
+                AuthenticationMetadata = model.AuthenticationMetadata
             });
 
             return CreateManagedInstanceModelFromResponse(resp);
@@ -299,6 +300,7 @@ namespace Microsoft.Azure.Commands.Sql.ManagedInstance.Adapter
             managedInstance.DatabaseFormat = resp.DatabaseFormat;
             managedInstance.PricingModel = resp.PricingModel;
             managedInstance.ExternalGovernanceStatus = resp.ExternalGovernanceStatus;
+            managedInstance.AuthenticationMetadata = resp.AuthenticationMetadata;
 
             return managedInstance;
         }