+++ src/Accounts/Accounts/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 
 ## Upcoming Release
+* Used Azure.Identity and Azure.Core directly for client assertion [#22628].
+
+## Version 3.0.3
 * Reduced the frequency of displaying sign-in announcement messages.
 * Upgraded Azure.Core to 1.41.0 to include the fix for `BearerTokenAuthenticationPolicy`
 * Removed the informational table about selected context to avoid duplication with output table.